Parser work

+++ parse.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
 import fire
 import pickle
+import progressbar
 
 from labm8 import fs
 from pycparser.plyparser import ParseError
-from pycparserext.ext_c_generator import OpenCLCGenerator
 from pycparserext.ext_c_parser import OpenCLCParser
 from random import shuffle
 
@@ -16,29 +16,34 @@ class Main(object):
         """
         indir = fs.path(indir)
         outdir = fs.path(outdir)
-
-        files = fs.ls(indir, abspaths=True)
-        n = len(files)
-        print("{n} input files".format(**vars()))
-
-        shuffle(files)
-
         fs.mkdir(outdir)
 
-        for i, path in enumerate(files):
-            if i % 10 == 0:
-                print("{i}/{n}".format(**vars()))
-            parser = OpenCLCParser()
-            with open(path) as infile:
-                src = infile.read()
+        # already done
+        done = set(fs.basename(x) for x in fs.ls(outdir))
+        # still to do
+        filenames = list(set(fs.basename(x) for x in fs.ls(indir)) - done)
+        # random order
+        shuffle(filenames)
 
-            try:
-                ast = parser.parse(src)
-                outpath = fs.path(outdir, fs.basename(path))
-                with open(outpath, "wb") as outfile:
-                    pickle.dump(ast, outfile)
-            except ParseError:
-                print(i + 1, "error", fs.basename(path))
+        print("{} input files".format(len(filenames)))
+
+        bar = progressbar.ProgressBar()
+        for filename in bar(filenames):
+            inpath = fs.path(indir, filename)
+            outpath = fs.path(outdir, filename)
+
+            if not fs.exists(outpath):
+                parser = OpenCLCParser()
+
+                with open(inpath) as infile:
+                    src = infile.read()
+
+                try:
+                    ast = parser.parse(src)
+                    with open(outpath, "wb") as outfile:
+                        pickle.dump(ast, outfile)
+                except ParseError:
+                    pass
 
 
 if __name__ == "__main__":
